+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 - Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
 
+## [0.4.0] - 2026-01-22
+### Added
+- Add Ice and Fire structure provider
+
+### Changed
+- Tweak the size ratios to better fit different screen resolutions
+
+
 ## [0.3.0] - 2026-01-10
 ### Added
 - Add AbyssalCraft structure provider